+++ src/modules/websocket/api/trade.js
@@ -199,7 +199,7 @@ const Trade = superclass => class extends superclass {
   }
 
   /**
-   * Place new OCO< br>
+   * Place new Order list< br>
    *
    * Send in a new OCO order.<br>
    *
@@ -207,33 +207,39 @@ const Trade = superclass => class extends superclass {
    *
    * @param {string} symbol
    * @param {string} side
-   * @param {number} price
    * @param {number} quantity
+   * @param {string} aboveType
+   * @param {string} belowType
    * @param {object} [options]
    * @param {string} [options.listClientOrderId]
-   * @param {string} [options.limitClientOrderId]
-   * @param {number} [options.limitIcebergQty]
-   * @param {number} [options.limitStrategyId]
-   * @param {number} [options.limitStrategyType]
-   * @param {number} [options.stopPrice]
-   * @param {number} [options.trailingDelta]
-   * @param {number} [options.stopClientOrderId]
-   * @param {number} [options.stopLimitPrice]
-   * @param {string} [options.stopLimitTimeInForce]
-   * @param {number} [options.stopIcebergQty]
-   * @param {number} [options.stopStrategyId]
-   * @param {string} [options.stopStrategyType]
-   * @param {number} [options.newOrderRespType]
-   * @param {number} [options.selfTradePreventionMode]
-   * @param {number} [options.recvWindow]
+   * @param {string} [options.aboveClientOrderId]
+   * @param {number} [options.aboveIcebergQty]
+   * @param {number} [options.abovePrice]
+   * @param {number} [options.aboveStopPrice]
+   * @param {number} [options.aboveTrailingDelta]
+   * @param {number} [options.aboveTimeInForce]
+   * @param {number} [options.aboveStrategyId]
+   * @param {number} [options.aboveStrategyType]
+   * @param {string} [options.belowClientOrderId]
+   * @param {number} [options.belowIcebergQty]
+   * @param {number} [options.belowPrice]
+   * @param {number} [options.belowStopPrice]
+   * @param {number} [options.belowTrailingDelta]
+   * @param {string} [options.belowTimeInForce]
+   * @param {number} [options.belowStrategyId]
+   * @param {number} [options.belowStrategyType]
+   * @param {string} [options.newOrderRespType]
+   * @param {string} [options.selfTradePreventionMode]
+   * @param {number} [options.recvWindow] - The value cannot be greater than 60000
    *
    */
-  newOCOOrder (symbol, side, price, quantity, options = {}) {
-    this.sendSignatureMessage('orderList.place', {
+  newOCOOrder (symbol, side, quantity, aboveType, belowType, options = {}) {
+    this.sendSignatureMessage('orderList.place.oco', {
       symbol,
       side,
-      price,
       quantity,
+      aboveType,
+      belowType,
       ...options
     })
   }